Move React Router hooks inside pages
Use React Router hooks directly inside each page instead of defining
wrappers around routes.
